pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "StraddlingCheckerboard"
dynamic = ["version"]
description = "A Flexible Digital version of the Straddling Checkerboard"
authors = [
    {name = "James David Clarke", email = "james@jamesdavidclarke.com"},
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "python-dotenv",
    "coloredlogs"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/JamesClarke7283/StraddlingCheckerboard"
"Bug Tracker" = "https://github.com/JamesClarke7283/StraddlingCheckerboard/issues"

[tool.setuptools]
packages = ["src"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["py.typed"]

[project.scripts]
straddling-checkerboard = "src.lib:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.setuptools_scm]